Remove `WaitForXYZ` trait abstraction
Modelling this as a trait unnecessarily couples it to a particular
connector implementations. Instead, we can create simple free-functions
for this that are actually an implementation detail of the herc20::new
function that no one outside of this module needs to be concerned with.
